requirements.txt
Pillow
PyQt6
matplotlib
numpy
pyqtgraph
pyvips
tifffile
torch
